challenge/medium_honey_i_shrunk_the_pets/requirements.txt
Pillow==8.4.0
torch==1.7.1
torchvision==0.8.2
